delete input event component and move to App
